Code Projects/Project 21/Market Research/relevance-market-intel/requirements.txt
python-dotenv>=1.0.1
relevanceai>=3.0.0
requests>=2.32.3
